ERM-3056, ERM-3057, ERM-3058, ERM-3059 Filter for documents in licenses and amendments
* avoid duplicate code
