+++ lib/DocumentFilter/DocumentFilter.js
@@ -14,7 +14,8 @@ import {
 
 import DocumentFilterForm from './DocumentFilterForm';
 
-const DocumentFilter = ({ activeFilters, filterHandlers }) => {
+const DocumentFilter = ({ activeFilters, atTypeValues = [], filterHandlers }) => {
+  const filterType = atTypeValues.length > 0 ? 'supplementaryDocuments' : 'documents';
   const [editingFilters, setEditingFilters] = useState(false);
   const openEditModal = () => setEditingFilters(true);
   const closeEditModal = () => setEditingFilters(false);
@@ -53,7 +54,7 @@ const DocumentFilter = ({ activeFilters, filterHandlers }) => {
     return [];
   };
 
-  const parsedFilterData = parseQueryString(activeFilters?.documents || []);
+  const parsedFilterData = parseQueryString(activeFilters?.[filterType] || []);
 
   const handleSubmit = (values) => {
     // In order to convert the form values into the shape for them to be deparsed we do the inverse of the above
@@ -82,7 +83,7 @@ const DocumentFilter = ({ activeFilters, filterHandlers }) => {
 
     filterHandlers.state({
       ...activeFilters,
-      documents: [
+      [filterType]: [
         // Currently the deparse function returns a query string containing whitespace which leads to grouping errors
         // This regex removes all whitespace from the querystring
         deparseKiwtQueryFilters(kiwtQueryShape).replaceAll(/\s/g, ''),
@@ -96,9 +97,9 @@ const DocumentFilter = ({ activeFilters, filterHandlers }) => {
       closedByDefault
       displayClearButton={!!parsedFilterData?.length}
       header={FilterAccordionHeader}
-      id="clickable-agreement-document-filter"
-      label={<FormattedMessage id="stripes-erm-components.documentFilter.coreDocuments" />}
-      onClearFilter={() => filterHandlers.state({ ...activeFilters, documents: [] })
+      id={`clickable-agreement-${filterType}-filter`}
+      label={<FormattedMessage id={`stripes-erm-components.documentFilter.${filterType}`} />}
+      onClearFilter={() => filterHandlers.state({ ...activeFilters, [filterType]: [] })
       }
       separator={false}
     >
@@ -111,6 +112,7 @@ const DocumentFilter = ({ activeFilters, filterHandlers }) => {
         </Layout>
       )}
       <DocumentFilterForm
+        atTypeValues={atTypeValues}
         editingFilters={editingFilters}
         filters={parsedFilterData}
         handlers={{
@@ -125,6 +127,7 @@ const DocumentFilter = ({ activeFilters, filterHandlers }) => {
 
 DocumentFilter.propTypes = {
   activeFilters: PropTypes.object,
+  atTypeValues: PropTypes.arrayOf(PropTypes.object),
   filterHandlers: PropTypes.object,
 };
