+++ lib/DocumentFilter/DocumentFilterField.js
@@ -3,7 +3,6 @@ import { Button, Row, Col, Label } from '@folio/stripes/components';
 import { useForm, useFormState } from 'react-final-form';
 import { FieldArray } from 'react-final-form-arrays';
 import { FormattedMessage } from 'react-intl';
-import SupplementaryDocumentFilterRule from './SupplementaryDocumentFilterRule';
 import DocumentFilterRule from './DocumentFilterRule';
 
 const DocumentFilterField = ({ atTypeValues, index, name }) => {
@@ -11,28 +10,17 @@ const DocumentFilterField = ({ atTypeValues, index, name }) => {
     mutators: { push },
   } = useForm();
   const { values } = useFormState();
-
   const renderRuleComponent = (ruleFields, ruleFieldName, ruleFieldIndex) => {
     return (
-      atTypeValues?.length > 0 ?
-        <SupplementaryDocumentFilterRule
-          key={ruleFieldName}
-          ariaLabelledby={`selected-document-item-name-${index}`}
-          atTypeValues={atTypeValues}
-          index={ruleFieldIndex}
-          name={ruleFieldName}
-          onDelete={() => ruleFields.remove(ruleFieldIndex)}
-          value={values.filters[index]?.rules[ruleFieldIndex]}
-        />
-        :
-        <DocumentFilterRule
-          key={ruleFieldName}
-          ariaLabelledby={`selected-document-item-name-${index}`}
-          index={ruleFieldIndex}
-          name={ruleFieldName}
-          onDelete={() => ruleFields.remove(ruleFieldIndex)}
-          value={values.filters[index]?.rules[ruleFieldIndex]}
-        />
+      <DocumentFilterRule
+        key={ruleFieldName}
+        ariaLabelledby={`selected-document-item-name-${index}`}
+        atTypeValues={atTypeValues}
+        index={ruleFieldIndex}
+        name={ruleFieldName}
+        onDelete={() => ruleFields.remove(ruleFieldIndex)}
+        value={values.filters[index]?.rules[ruleFieldIndex]}
+      />
     );
   };
