+++ lib/DocumentFilter/DocumentFilterRule.js
@@ -10,20 +10,22 @@ import {
   Select,
   TextField,
 } from '@folio/stripes/components';
-import { FormattedMessage } from 'react-intl';
+import { FormattedMessage, useIntl } from 'react-intl';
 import { Field } from 'react-final-form';
 import { required as requiredValidator } from '../validators';
 import DocumentFilterRuleConstants from './DocumentFilterRuleConstants';
 
 const DocumentFilterRule = ({
   ariaLabelledby,
+  atTypeValues,
   index,
   name,
   onDelete,
   value,
 }) => {
-  const { operators, docs: attributes } = DocumentFilterRuleConstants();
-
+  const intl = useIntl();
+  // atTypeValues.length > 0
+  const { operators, attributes } = DocumentFilterRuleConstants(atTypeValues.length > 0);
   return (
     <Row key={name}>
       <Col xs={2}>
@@ -50,7 +52,25 @@ const DocumentFilterRule = ({
             <Select
               {...input}
               aria-labelledby={`${ariaLabelledby}-rule-column-header-comparator`}
-              dataOptions={operators}
+              dataOptions={
+                value?.path === 'supplementaryDocs.atType.value'
+                  ? [
+                    { label: '', value: '' },
+                    {
+                      label: intl.formatMessage({
+                        id: 'stripes-erm-components.operator.is',
+                      }),
+                      value: '==',
+                    },
+                    {
+                      label: intl.formatMessage({
+                        id: 'stripes-erm-components.operator.isNot',
+                      }),
+                      value: '!=',
+                    },
+                  ]
+                  : operators
+              }
               error={meta?.touched && meta?.error}
               required
             />
@@ -59,13 +79,26 @@ const DocumentFilterRule = ({
       </Col>
       <Col xs={3}>
         <Field name={`${name}.value`} validate={requiredValidator}>
-          {({ input }) => (
-            <TextField
-              {...input}
-              aria-labelledby={`${ariaLabelledby}-rule-column-header-value`}
-              disabled={!value?.path}
-              required
-            />
+          {({ input, meta }) => (
+            <>
+              {value?.path === 'supplementaryDocs.atType.value' ? (
+                <Select
+                  {...input}
+                  aria-labelledby={`${ariaLabelledby}-rule-column-header-value`}
+                  dataOptions={[{ label: '', value: '' }, ...atTypeValues]}
+                  disabled={!value?.path}
+                  error={meta?.touched && meta?.error}
+                  required
+                />
+              ) : (
+                <TextField
+                  {...input}
+                  aria-labelledby={`${ariaLabelledby}-rule-column-header-value`}
+                  disabled={!value?.path}
+                  required
+                />
+              )}
+            </>
           )}
         </Field>
       </Col>
@@ -97,6 +130,7 @@ const DocumentFilterRule = ({
 
 DocumentFilterRule.propTypes = {
   ariaLabelledby: PropTypes.string.isRequired,
+  atTypeValues: PropTypes.arrayOf(PropTypes.object),
   index: PropTypes.number.isRequired,
   name: PropTypes.string.isRequired,
   onDelete: PropTypes.func.isRequired,
