+++ lib/DocumentFilter/DocumentFilterRuleConstants.js
@@ -1,6 +1,6 @@
 import { useIntl } from 'react-intl';
 
-const DocumentFilterRuleConstants = () => {
+const DocumentFilterRuleConstants = (useSupplementaryDocs) => {
   const intl = useIntl();
 
   const operators = [
@@ -78,7 +78,9 @@ const DocumentFilterRuleConstants = () => {
     },
   ];
 
-  return { docs, operators, supplementaryDocs };
+  const attributes = useSupplementaryDocs ? supplementaryDocs : docs;
+
+  return { operators, attributes };
 };
 
 export default DocumentFilterRuleConstants;
