+++ lib/DocumentFilter/index.js
@@ -1,2 +1,1 @@
-export { default as DocumentFilter } from './DocumentFilter';
-export { default as SupplementaryDocumentFilter } from './SupplementaryDocumentFilter';
+export { default } from './DocumentFilter';
