+++ translations/stripes-erm-components/en.json
@@ -203,8 +203,7 @@
   "doc.error.docsMustHaveLocationOrURL": "Each document must specify a file, location, or URL.",
   "doc.error.invalidURL": "Please enter a valid URL (including \"http\" or \"https\").",
   "doc.error.invalidSize": "URL may only be {size} characters long",
-  "documentFilter.documents": "Documents",
-  "documentFilter.coreDocuments": "Core documents",
+  "documentFilter.documents": "Core documents",
   "documentFilter.supplementaryDocuments": "Supplementary documents",
   "documentFilter.filtersApplied": "{filtersLength} document filters applied",
   "documentFilter.editDocumentFilters": "Edit document filters",
